Add `useFocusCapture` hook
Factor out a common `useKeyboardNavigation` hook, used by existing
`useArrowKeyNavigation` hook and new `useFocusCapture`.
